feat: adaptor of go-ethereum v1.13

+++ storage/pebble/pebble_test.go
@@ -426,7 +426,7 @@ func BenchmarkPebbleSuite(b *testing.B) {
 
 		// The size of memory table(as well as the write buffer).
 		// Note, there may have more than two memory tables in the system.
-		MemTableSize: memTableSize,
+		MemTableSize: uint64(memTableSize),
 
 		// MemTableStopWritesThreshold places a hard limit on the size
 		// of the existent MemTables(including the frozen one).
